Small refactor ahead of bigger work to keep project structure clean
